% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignNames.R
\name{assignNames}
\alias{assignNames}
\title{assignNames}
\usage{
assignNames(x)
}
\arguments{
\item{x}{list}
}
\description{
Recursively assigns names to the children of a list of lists.
}
\author{
Alex Howard
}
