% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenienceFunctions.R
\name{extractTopic}
\alias{extractTopic}
\title{extractTopic
Convenience function to extract topicId from policy}
\usage{
extractTopic(policy)
}
\arguments{
\item{policy}{}
}
\description{
extractTopic
Convenience function to extract topicId from policy
}
\examples{
extractTopic(policy)
}
