% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GassociateUserToken.R
\name{GassociateUserToken}
\alias{GassociateUserToken}
\title{GassociateUserToken}
\usage{
GassociateUserToken(refreshToken, tokenId, baseurl = "http://localhost:3000/")
}
\arguments{
\item{refreshToken}{Character. JWT refresh token returned by Glogin()$refreshToken.}

\item{tokenId}{Character. The ID of the token, e.g., "0.0.2950835".}

\item{baseurl}{Character. URL where /api/ is hosted. Default "http://localhost:3000/".}
}
\value{
Logical. TRUE if successful.
}
\description{
Associates a Guardian user with a token of a Guardian policy.
}
