% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GcreateUser.R
\name{GcreateUser}
\alias{GcreateUser}
\title{GcreateUser}
\usage{
GcreateUser(
  un,
  pw,
  role = "USER",
  parentDid,
  hederaAccountId,
  hederaAccountKey,
  baseurl = "http://localhost:3000/"
)
}
\arguments{
\item{un}{Character. The username to be registered.}

\item{pw}{Character. The password to be registered.}

\item{role}{Character. The Guardian user type to be registered. Defaults to "USER".}

\item{parentDid}{Character. The DID of the standard registry user to associate the user with.}

\item{hederaAccountId}{Character. The Hedera account ID to be associated with the user.}

\item{hederaAccountKey}{Character. The Hedera account key to be associated with the user.}

\item{baseurl}{Character. URL where /api/ is hosted. Default "http://localhost:3000/".}
}
\value{
A list with at least the following elements:
 - id: The user's unique Guardian identifier (e.g., "65ca39498ae968bb194096be")
 - username: The user's Guardian username (e.g., "jdoe")
 - hash_password: The user's hashed Guardian password (e.g., "65ca39498ae968bb194096be")
 - role: The user's role (e.g., "STANDARD_REGISTRY or "USER")
 - refreshToken: The user's JWT refresh token.
 - did: The user's DID.
 - didDocMessageId: The message ID of the user's DID document.
}
\description{
Convenience wrapper around the GregisterAccount, Glogin and
   GsetUpProfile functions which should almost always be called in succession.
}
