% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GgetAccessToken.R
\name{GgetAccessToken}
\alias{GgetAccessToken}
\title{GgetAccessToken}
\usage{
GgetAccessToken(refreshToken, baseurl = "http://localhost:3000/")
}
\arguments{
\item{refreshToken}{Character. JWT refresh token returned by Glogin()$refreshToken.}

\item{baseurl}{Character. URL where /api/ is hosted. Default "http://localhost:3000/".}
}
\value{
Character. A JWT access token valid for 30 seconds.
}
\description{
GgetAccessToken
}
