% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GgetAccounts.R
\name{GgetAccounts}
\alias{GgetAccounts}
\title{GgetAccounts}
\usage{
GgetAccounts(refreshToken, baseurl = "http://localhost:3000")
}
\arguments{
\item{refreshToken}{Character. JWT refresh token returned by Glogin()$refreshToken.}

\item{baseurl}{Character. Base URL of the targeted Guardian instance. Defaults
to "http://localhost:3000/".}
}
\value{
Data frame.
}
\description{
Returns all users except those with roles Standard Registry and
  Auditor. Only users with the Standard Registry role are allowed to make the
  request.
}
