% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GgetBlockDocuments.R
\name{GgetBlockDocuments}
\alias{GgetBlockDocuments}
\title{GgetBlockDocuments}
\usage{
GgetBlockDocuments(
  refreshToken,
  policyId,
  blockId,
  baseurl = "http://localhost:3000/",
  inclFileMd = TRUE
)
}
\arguments{
\item{refreshToken}{Character. JWT refresh token returned by Glogin()$refreshToken.}

\item{policyId}{Character. ID of the policy from which to retrieve the documents.}

\item{blockId}{Character. ID of the block from which to retrieve the documents.}

\item{baseurl}{Character. URL where /api/ is hosted. Default "http://localhost:3000/".}

\item{inclFileMd}{Logical. If TRUE, file metadata is included in the returned
list. Defaults to TRUE.}
}
\value{
List. List of documents associated with the block.
}
\description{
GgetBlockDocuments
}
