% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GgetHttpRequestBlocks.R
\name{GgetHttpRequestBlocks}
\alias{GgetHttpRequestBlocks}
\title{GgetHttpRequestBlocks}
\usage{
GgetHttpRequestBlocks(
  refreshToken = NULL,
  baseurl = "http://localhost:3000/",
  policyId = NULL
)
}
\arguments{
\item{refreshToken}{Character . A valid refreshtoken}

\item{baseurl}{Character. A valid Url. Defauklt "http://localhost:3000/"}

\item{policyId}{Character. A Guardian policy ID}
}
\value{
tibble
}
\description{
Convenience function to get just the httpRequestBlock from a policy
}
\examples{
GgetHttpRequestBlocks("65c3cbe7320d203db7dd6e44")
}
