% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GgetPolicies.R
\name{GgetPolicies}
\alias{GgetPolicies}
\title{GgetPolicies}
\usage{
GgetPolicies(
  refreshToken,
  baseurl = "http://localhost:3000/",
  pageIndex = 0,
  pageSize = 10,
  returndf = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{refreshToken}{Character. JWT refresh token returned by Glogin()$refreshToken.}

\item{baseurl}{Character. Base url. defaults to "http://localhost:3000/"}

\item{pageIndex}{Integer. default 0}

\item{pageSize}{Integer. default 10}

\item{returndf}{Logical. Returns a tibble if TRUE. Else a list (default)}
}
\value{
tibble or list
}
\description{
Get the policies associated with an user identified by an access token or username and passwork
}
