% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GgetPolicyBlocks.R
\name{GgetPolicyBlocks}
\alias{GgetPolicyBlocks}
\title{GgetPolicyBlocks}
\usage{
GgetPolicyBlocks(
  refreshToken = NULL,
  baseurl = "http://localhost:3000/",
  policyId = NULL,
  returndf = FALSE,
  schemaBlocksOnly = TRUE
)
}
\arguments{
\item{refreshToken}{Character. JWT refresh token returned by Glogin()$refreshToken.}

\item{baseurl}{Character. Base url. defaults to "http://localhost:3000/"}

\item{policyId}{Character. Typically from the $id item of the result of GgetPolicies()}

\item{schemaBlocksOnly}{Logical. If TRUE, only return blocks that have a schema.}
}
\value{
tibble
}
\description{
GgetPolicyBlocks
}
