% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GgetPolicyConfig.R
\name{GgetPolicyConfig}
\alias{GgetPolicyConfig}
\title{Title}
\usage{
GgetPolicyConfig(
  refreshToken = NULL,
  baseurl = "http://localhost:3000/",
  policyId = NULL,
  returndf = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{refreshToken}{Character. JWT refresh token returned by Glogin()$refreshToken.}

\item{baseurl}{Character. Base url. defaults to "http://localhost:3000/"}

\item{policyId}{haracter. Typically from the $id item of the result of GgetPolicies()}

\item{returndf}{Return a tibble or not Default FALSE.}

\item{...}{}
}
\description{
Title
}
