% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GgetPolicyMessageId.R
\name{GgetPolicyMessageId}
\alias{GgetPolicyMessageId}
\title{GgetPolicyMessageId}
\usage{
GgetPolicyMessageId(policyId, refreshToken, baseurl = "http://localhost:3000/")
}
\arguments{
\item{policyId}{The ID of the policy.}

\item{refreshToken}{Character. JWT refresh token returned by Glogin()$refreshToken.}

\item{baseurl}{Character. Base url of the targeted Guardian instance.
Defaults to "http://localhost:3000/".}
}
\value{
Character. The message ID of the policy.
}
\description{
Gets the message ID of a policy.
}
