% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GgetPolicyTokens.R
\name{GgetPolicyTokens}
\alias{GgetPolicyTokens}
\title{GgetPolicyTokens}
\usage{
GgetPolicyTokens(refreshToken, policyId, baseurl = "http://localhost:3000/")
}
\arguments{
\item{refreshToken}{Character. JWT refresh token returned by Glogin()$refreshToken.}

\item{policyId}{Character. Policy ID to get token info for.}

\item{baseurl}{Character. Base url. Defaults to "http://localhost:3000/".}
}
\value{
Data frame.
}
\description{
Get a list of tokens in use in a specific policy.
}
