% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GgetSchemas.R
\name{GgetSchemas}
\alias{GgetSchemas}
\title{GgetSchemas}
\usage{
GgetSchemas(
  accessToken = NULL,
  baseurl = "http://localhost:3000/",
  un = NULL,
  pw = NULL,
  pageIndex = 0,
  pageSize = 10,
  topicId = NULL,
  schemaId = NULL,
  returndf = FALSE
)
}
\arguments{
\item{accessToken}{Character. Access token returned by Glogin()  $accessToken}

\item{baseurl}{Character. Base url. defaults to "http://localhost:3000/"}

\item{un}{Character. Username}

\item{pw}{haracter. Password}

\item{pageIndex}{Integer. default 0}

\item{pageSize}{Integer. default 10}

\item{topicId}{Character}
}
\description{
Get schemas based on a topicID
}
