% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GgetStandardsRegistries.R
\name{GgetStandardsRegistries}
\alias{GgetStandardsRegistries}
\title{GgetStandardsRegistries}
\usage{
GgetStandardsRegistries(refreshToken, baseurl = "http://localhost:3000/")
}
\arguments{
\item{refreshToken}{Character. JWT refresh token returned by Glogin()$refreshToken.}

\item{baseurl}{Character. Base url. Defaults to "http://localhost:3000/".}
}
\value{
Data frame.
}
\description{
Get the names and DIDs of all standards registry users on the Guardian instance at <baseurl>.
}
