% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GgetTaskStatus.R
\name{GgetTaskStatus}
\alias{GgetTaskStatus}
\title{GgetTaskStatus}
\usage{
GgetTaskStatus(taskId, refreshToken, baseurl = "http://localhost:3000/")
}
\arguments{
\item{taskId}{Character. The ID of the task to get the status of.}

\item{refreshToken}{Character. JWT refresh token returned by Glogin()$refreshToken.}

\item{baseurl}{Character. Base URL of the targeted Guardian instance. Defaults
to "http://localhost:3000/".}
}
\value{
A list with the task details and status.
}
\description{
Returns the status of a Guardian task. Only users with the Standard
  Registry role are allowed to make the request.
}
