% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GgetUserTokenInfo.R
\name{GgetUserTokenInfo}
\alias{GgetUserTokenInfo}
\title{GgetUserTokenInfo}
\usage{
GgetUserTokenInfo(
  refreshToken,
  tokenId,
  usernm,
  baseurl = "http://localhost:3000/"
)
}
\arguments{
\item{refreshToken}{Character. JWT refresh token returned by Glogin()$refreshToken.}

\item{tokenId}{Character. The ID of the token of interest, e.g., "0.0.7717332".}

\item{usernm}{Character. The username of the user of interest.}

\item{baseurl}{Character. Base URL of the targeted Guardian instance. Defaults
to "http://localhost:3000/".}
}
\value{
Data frame.
}
\description{
Returns user information for the selected token. Only users with
   the Standard Registry role are allowed to make the request.
}
