% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Glist2tibble.R
\name{Glist2tibble}
\alias{Glist2tibble}
\title{Glist2tibble}
\usage{
Glist2tibble(L, flatfirst = FALSE)
}
\arguments{
\item{L}{list}

\item{flatfirst}{logical}
}
\value{
tibble
}
\description{
Glist2tibble
}
