% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Glogin.R
\name{Glogin}
\alias{Glogin}
\title{Glogin}
\usage{
Glogin(
  un = Sys.getenv("guardianUn"),
  pw = Sys.getenv("guardianPW"),
  baseurl = "http://localhost:3000/"
)
}
\arguments{
\item{un}{Character. Username}

\item{pw}{Character. Password}

\item{baseurl}{Character. Base url. defaults to "http://localhost:3000/"}
}
\value{
list with "username", "did", "role" and "accessToken"
}
\description{
Glogin
}
