% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GmakeBody.R
\name{GmakeBody}
\alias{GmakeBody}
\title{GmakeBody}
\usage{
GmakeBody(document = NULL, schemaID = NULL, schemaIPFSurl = NULL)
}
\arguments{
\item{document}{Named list or data frame with data to submit. Names must correspond to field names in Schema}

\item{schemaID}{Character}

\item{schemaIPFSurl}{Character}
}
\description{
GmakeBody
}
