% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GmakePolicyCopy.R
\name{GmakePolicyCopy}
\alias{GmakePolicyCopy}
\title{GmakePolicyCopy}
\usage{
GmakePolicyCopy(
  refreshToken,
  messageId = NULL,
  fp = NULL,
  newName = NULL,
  newVersionNo = NULL,
  publish = TRUE,
  baseurl = "http://localhost:3000/"
)
}
\arguments{
\item{refreshToken}{Character. JWT refresh token returned by Glogin()$refreshToken.}

\item{messageId}{Character. The message ID of the policy to be copied. If provided,
the function will ignore the 'fp' argument. If not provided, the function will
try to copy the policy from the ZIP file provided in the 'fp' argument.}

\item{fp}{Character. The file path of the ZIP file of the policy to be copied.
Ignored if 'messageId' is provided.}

\item{newName}{Character. Optional. The name to give the new copy of the policy.}

\item{newVersionNo}{Character. Optional. The new version number of the policy. If
not provided and publish = TRUE, the function will try to determine the new
version number automatically. It is strongly recommended to provide the correct
version number yourself. This argument is ignored if publish = FALSE.}

\item{publish}{Logical. Whether to publish the new policy. Defaults to TRUE.}

\item{baseurl}{Character. Base URL of the targeted Guardian instance. Defaults
to "http://localhost:3000/".}
}
\value{
Character. The Guardian policy ID of the newly created copy.
}
\description{
Makes a copy of a Guardian policy and pushes it back to the
   Guardian instance. Can optionally change the name of the copy and publish it.
}
