% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GmakeSchemaTemplate.R
\name{GmakeSchemaTemplate}
\alias{GmakeSchemaTemplate}
\title{GmakeSchemaTemplate}
\usage{
GmakeSchemaTemplate(dfSchemas)
}
\arguments{
\item{dfSchemas}{tibble. Output of Gschema2tibble()}
}
\description{
Make a nested tibble by schema. The data fields and their types are in the nested column
}
