% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gpost2block.R
\name{Gpost2block}
\alias{Gpost2block}
\title{Gpost2block}
\usage{
Gpost2block(
  refreshToken,
  document,
  schemaID,
  schemaIPFSurl,
  policyId,
  BlockId,
  baseurl = "http://localhost:3000/"
)
}
\arguments{
\item{refreshToken}{Character. JWT refresh token returned by Glogin()$refreshToken.}

\item{document}{Named list or data frame with data to submit. Names must correspond to field names in Schema.}

\item{schemaID}{character. Schema to submit to.}

\item{schemaIPFSurl}{character. Schema to submit to.}

\item{policyId}{character. Schema to submit to.}

\item{BlockId}{character. Schema to submit to.}

\item{baseurl}{Character. URL where /api/ is hosted. Default "http://localhost:3000/".}
}
\value{
NULL. #TODO. This is not accurate. It does return stuff.
}
\description{
Gpost2block
}
