% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gpost2Schema.R
\name{Gpost2Schema}
\alias{Gpost2Schema}
\title{Gpost2Schema}
\usage{
Gpost2Schema(
  accessToken = NULL,
  document = NULL,
  schemaID = NULL,
  schemaIPFSurl = NULL,
  policyId = NULL,
  BlockId = NULL,
  baseurl = "http://localhost:3000/"
)
}
\arguments{
\item{accessToken}{character. Access token. typically from 'accessToken' field in result of Glogin()}

\item{document}{Named list or data frame with data to submit. Names must correspond to field names in Schema}

\item{schemaID}{character. Schema to submit to}

\item{schemaIPFSurl}{character. Schema to submit to}

\item{policyId}{character. Schema to submit to}

\item{BlockId}{character. Schema to submit to}

\item{baseurl}{Character. URL where /api/ is hosted. Default "http://localhost:3000/"}
}
\description{
Gpost2Schema
}
