% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GpublishPolicy.R
\name{GpublishPolicy}
\alias{GpublishPolicy}
\title{GpublishPolicy}
\usage{
GpublishPolicy(
  refreshToken,
  policyId,
  newVersionNo = NULL,
  baseurl = "http://localhost:3000/"
)
}
\arguments{
\item{refreshToken}{Character. JWT refresh token returned by Glogin()$refreshToken.}

\item{policyId}{Character. The Guardian ID of the draft policy to rename.}

\item{newVersionNo}{Character. Optional. The new version number of the policy. If
not provided and publish = TRUE, the function will try to determine the new
version number automatically. It is strongly recommended to provide the correct
version number yourself. This argument is ignored if publish = FALSE.}

\item{baseurl}{Character. Base URL of the targeted Guardian instance. Defaults
to "http://localhost:3000/".}
}
\value{
Logical. TRUE if the policy was successfully published; will throw
  an error otherwise.
}
\description{
Changes a draft policy's status to 'PUBLISH' in the Guardian.
}
