% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GregisterAccount.R
\name{GregisterAccount}
\alias{GregisterAccount}
\title{GregisterAccount}
\usage{
GregisterAccount(un, pw, role = "USER", baseurl = "http://localhost:3000/")
}
\arguments{
\item{un}{Character. The username to be registered.}

\item{pw}{Character. The password to be registered.}

\item{role}{Character. The Guardian user type to be registered. Defaults to "USER".}

\item{baseurl}{Character. URL where /api/ is hosted. Default "http://localhost:3000/".}
}
\value{
A list with at least the following elements:
 - id: The user's unique Guardian identifier (e.g., "65ca39498ae968bb194096be")
 - username: The user's username (e.g., "jdoe")
 - role: The user's role (e.g., "STANDARD_REGISTRY or "USER")
}
\description{
Creates a new Guardian user account on the Guardian instance at {baseurl}.
   Note that this function does not associate the new user with any specific
   standards registry on the Guardian instance in question; the association
   with a standards registry must be done separately.
}
