% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gschema2tibble.R
\name{Gschema2tibble}
\alias{Gschema2tibble}
\title{Gschema2tibble}
\usage{
Gschema2tibble(schemas)
}
\arguments{
\item{schemas}{list}
}
\value{
tibble
}
\description{
take the list comming out of GgetSchemas and make a tibble with 1 schema detail per row
}
