% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GsetUpProfile.R
\name{GsetUpProfile}
\alias{GsetUpProfile}
\title{GsetUpProfile}
\usage{
GsetUpProfile(
  refreshToken,
  un,
  role,
  parentDid,
  hederaAccountId,
  hederaAccountKey,
  baseurl = "http://localhost:3000/"
)
}
\arguments{
\item{refreshToken}{Character. JWT refresh token returned by Glogin()$refreshToken.}

\item{un}{Character. The username to be registered.}

\item{role}{Character. The Guardian user type to be registered. Defaults to "USER".}

\item{parentDid}{Character. The DID of the standard registry user to associate the user with.}

\item{hederaAccountId}{Character. The Hedera account ID to be associated with the user.}

\item{hederaAccountKey}{Character. The Hedera account key to be associated with the user.}

\item{baseurl}{Character. Base URL of the targeted Guardian instance. Defaults
to "http://localhost:3000/".}
}
\description{
Associates a Guardian user with a Hedera account and a Guardian
 standard registry user.
}
