% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name2tibble.R
\name{names2tibble}
\alias{names2tibble}
\title{names2tibble}
\usage{
names2tibble(data, fieldnames = "fieldnames")
}
\arguments{
\item{data}{a dataframe with a column with fieldnames}

\item{fieldnames}{Character. names of columns to be created}
}
\description{
names2tibble
}
