% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nestedSchemaTemplate.R
\name{nestedSchemaTemplate}
\alias{nestedSchemaTemplate}
\title{nestedSchemaTemplate}
\usage{
nestedSchemaTemplate(dfSchemas = NULL, id = NULL, uuid = "uuid")
}
\arguments{
\item{dfSchemas}{tibble.  Output of Gschema2tibble()}

\item{id}{character. id field in dfSchema of schema for which you want to make the template}
}
\description{
Make a template for a schema with sub-schema field nested in list-columns
}
