% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenienceFunctions.R
\name{ones}
\alias{ones}
\title{ones
Convenience function to extract names of columns that contain nested items of length 1}
\usage{
ones(df)
}
\arguments{
\item{df}{}
}
\value{
character vector with names of df that contain nested items of length 1
}
\description{
ones
Convenience function to extract names of columns that contain nested items of length 1
}
\examples{
df \%>\% unnest(cols = ones(.))
}
