% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whichBlocksHasSchema.R
\name{whichBlocksHasSchema}
\alias{whichBlocksHasSchema}
\title{whichBlocksHasSchema}
\usage{
whichBlocksHasSchema(
  dfBlocks,
  schemaId,
  washfunction = c(function(x) {
     gsub("#", "", x)
 }, NULL)[1],
  schemaname = "schema"
)
}
\arguments{
\item{dfBlocks}{tibble. Typically the output of GgetPolicyBlocks()}

\item{schemaId}{character. The uuid of a schema}

\item{washfunction}{function or NULL. Function to clean the variable containing the schemaId (specified by schemaname)
in dfBlocks. Default function(x){gsub("#", "", x)} . If NULL no cleaing is performed}

\item{schemaname}{Character. Variable in dfBlocks in which to look for schemaId}
}
\value{
tibble
}
\description{
Check in a collection of blocks which blocks are associated with a particular schema
}
\examples{
b <- whichBlocksHasSchema(GgetPolicyBlocks(AT, policyId = policyid),  "a399cc58-9a79-4277-b5ef-1c6461843a58")
}
