% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenienceFunctions.R
\name{extractTopic}
\alias{extractTopic}
\title{extractTopic}
\usage{
extractTopic(policy)
}
\arguments{
\item{policy}{Character.}
}
\description{
Convenience function to extract topicId from policy.
}
