% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GgetPolicies.R
\name{GgetPolicies}
\alias{GgetPolicies}
\title{GgetPolicies}
\usage{
GgetPolicies(
  accessToken = NULL,
  un = NULL,
  pw = NULL,
  baseurl = "http://localhost:3000/",
  pageIndex = 0,
  pageSize = 10,
  returndf = FALSE
)
}
\arguments{
\item{accessToken}{Character. Access token returned by Glogin()  $accessToken}

\item{un}{Character. Username}

\item{pw}{Character. Password}

\item{baseurl}{Character. Base url. defaults to "http://localhost:3000/"}

\item{pageIndex}{Integer. default 0}

\item{pageSize}{Integer. default 10}

\item{returndf}{Logical. Returns a tibble if TRUE. Else a list (default)}
}
\description{
Get the policies associated with an user identified by an access token or username and passwork
}
