% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GgetPolicy.R
\name{GgetPolicy}
\alias{GgetPolicy}
\title{GgetPolicy}
\usage{
GgetPolicy(
  refreshToken,
  baseurl = "http://localhost:3000/",
  policyId = NULL,
  returndf = FALSE,
  assignNms = TRUE
)
}
\arguments{
\item{refreshToken}{Character. JWT refresh token returned by Glogin()$refreshToken.}

\item{baseurl}{Character. Base url. defaults to "http://localhost:3000/"}

\item{policyId}{Character. Typically from the $id item of the result of GgetPolicies()}

\item{returndf}{Return a tibble or not Default FALSE.}
}
\value{
tibble or list
}
\description{
GgetPolicy
}
