% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GgetPolicyBlocks.R
\name{GgetPolicyBlocks}
\alias{GgetPolicyBlocks}
\title{GgetPolicyBlocks}
\usage{
GgetPolicyBlocks(
  accessToken = NULL,
  un = NULL,
  pw = NULL,
  baseurl = "http://localhost:3000/",
  policyId = NULL,
  returndf = FALSE
)
}
\arguments{
\item{accessToken}{Character. Access token returned by Glogin()  $accessToken}

\item{un}{Character. Username}

\item{pw}{Character. Password}

\item{baseurl}{Character. Base url. defaults to "http://localhost:3000/"}

\item{policyId}{Character. Typically from the $id item of the result of GgetPolicies()}
}
\value{
tibble
}
\description{
GgetPolicyBlocks
}
