% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GgetSchemas.R
\name{GgetSchemas}
\alias{GgetSchemas}
\title{GgetSchemas}
\usage{
GgetSchemas(
  refreshToken,
  baseurl = "http://localhost:3000/",
  pageIndex = 0,
  pageSize = 10,
  topicId = NULL,
  schemaId = NULL,
  returndf = FALSE
)
}
\arguments{
\item{refreshToken}{Character. JWT refresh token returned by Glogin()$refreshToken.}

\item{baseurl}{Character. Base url. defaults to "http://localhost:3000/"}

\item{pageIndex}{Integer. default 0}

\item{pageSize}{Integer. default 10}

\item{topicId}{Character}
}
\value{
tibble or list
}
\description{
Get schemas based on a topicID
}
