% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GgetUserProfile.R
\name{GgetUserProfile}
\alias{GgetUserProfile}
\title{GgetUserProfile}
\usage{
GgetUserProfile(refreshToken, username, baseurl = "http://localhost:3000/")
}
\arguments{
\item{refreshToken}{Character. JWT refresh token returned by Glogin()$refreshToken.}

\item{username}{Character. Username of the user for whom to retrieve the profile.}

\item{baseurl}{Character. Base url. Defaults to "http://localhost:3000/".}
}
\value{
Data frame.
}
\description{
Get user profile.
}
