% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gpost2block.R
\name{Gpost2block}
\alias{Gpost2block}
\title{Gpost2block}
\usage{
Gpost2block(
  x,
  refreshToken,
  policyId,
  blockId,
  baseurl = "http://localhost:3000/"
)
}
\arguments{
\item{x}{List. Data to submit to the policy block. If x is a document based
on a schema, the document should already contain appropriately populated
'context' and 'type' fields.}

\item{refreshToken}{Character. JWT refresh token returned by Glogin()$refreshToken.}

\item{policyId}{character. Schema to submit to.}

\item{blockId}{character. Schema to submit to.}

\item{baseurl}{Character. URL where /api/ is hosted. Default "http://localhost:3000/".}
}
\value{
NULL. #TODO. This is not accurate. It does return stuff.
}
\description{
Gpost2block
}
