% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenienceFunctions.R
\name{ones}
\alias{ones}
\title{ones}
\usage{
ones(df)
}
\value{
Character vector with names of df that contain nested items of length 1.
}
\description{
Convenience function to extract names of columns that contain nested items of length 1.
}
